+++ apps/www/CHANGELOG.md
@@ -1,5 +1,14 @@
 # www
 
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @repo/ai@2.0.1
+  - @repo/contents@1.0.1
+  - @repo/backend@2.0.1
+
 ## 2.0.0
 
 ### Major Changes